agent/pyproject.toml
[project]
name = "starnion-agent"
version = "0.1.0"
description = "Starnion AI Agent - Personal assistant powered by LangGraph and Gemini"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langgraph>=0.4",
    "langchain-google-genai>=2.1",
    "langchain-anthropic>=0.3",
    "langchain-openai>=0.3",
    "langgraph-checkpoint-postgres>=2.0",
    "psycopg[binary]>=3.2",
    "psycopg-pool>=3.2",
    "pydantic>=2.10",
    "pyyaml>=6.0",
    "grpcio>=1.70",
    "grpcio-tools>=1.70",
    "google-genai>=1.0",
    "pypdf>=5.0",
    "pillow>=11.0",
    "httpx>=0.28",
    "python-docx>=1.1",
    "openpyxl>=3.1",
    "python-pptx>=1.0",
    "olefile>=0.47",
    "google-api-python-client>=2.150",
    "google-auth-oauthlib>=1.2",
    "google-auth-httplib2>=0.2",
    "reportlab>=4.4.10",
    "tavily-python>=0.5",
    "readability-lxml>=0.8",
    "chardet<6",
    "qrcode[pil]>=8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.25",
    "ruff>=0.9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/starnion_agent"]
artifacts = ["src/starnion_agent/document/fonts/*.ttf"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py313"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[dependency-groups]
dev = [
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]
